requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
bitsandbytes==0.44.1
bleach==6.2.0
blinker==1.9.0
blis==1.0.1
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
compressed-tensors==0.8.0
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.8
datasets==3.2.0
decorator==5.1.1
decord==0.6.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
executing==2.1.0
fairscale==0.4.4
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.13.1
Flask==3.1.0
fonttools==4.55.0
frozenlist==1.5.0
fschat==0.2.36
fsspec==2024.2.0
ftfy==6.3.1
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.0
huggingface-hub==0.26.2
identify==2.6.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
interegular==0.3.3
iopath==0.1.10
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
jiter==0.8.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lark==1.2.2
latex2mathml==3.77.0
lazy_loader==0.4
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
llvmlite==0.43.0
lm-format-enforcer==0.10.9
lmdb==1.5.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
narwhals==1.13.5
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.19
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==2.29.0
openai==1.55.3
opencv-python==4.10.0.84
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.10.12
outlines==0.0.46
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post4
peft==0.13.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==2.10.1
pre_commit==4.0.1
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.18
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
ruff==0.8.1
safetensors==0.4.5
salesforce-lavis==1.0.2
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.18.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.3
streamlit==1.40.1
svgwrite==1.4.3
sympy==1.13.1
tenacity==9.0.0
text-unidecode==1.3
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.7.0
timm==0.6.13
tokenizers==0.15.1
toml==0.10.2
tomlkit==0.12.0
torch==2.1.2
torchaudio==2.1.0+cu121
torchvision==0.16.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.37.2
triton==2.1.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.27.1
vllm==0.6.4.post1
wandb==0.18.7
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.16.0
xformers==0.0.28.post3
xxhash==3.5.0
yarl==1.17.1
zipp==3.21.0
